requirements.txt
altgraph==0.17.2
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.11
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-jazzmin==2.4.9
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
future==0.18.2
idna==3.3
mysqlclient==2.1.0
pefile==2021.9.3
phonenumbers==8.12.45
Pillow==9.0.1
pyinstaller-hooks-contrib==2022.2
PyJWT==2.3.0
pytz==2021.3
pywin32-ctypes==0.2.0
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0
